refactor: redux 적용, fix: 글 작성 가능
