+++ client/src/components/AuthPostList.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import '../styles/AuthPostList.css';
 import { getAuthPosts } from 'api/api.js';
-import { Interface } from 'readline';
+// import { Interface } from 'readline';
 
 const AuthPostList = () => {
   const [allAuthPosts, setAllAuthPosts] = useState([]);
