+++ client/src/store/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from "@reduxjs/toolkit";
+import rootReducer from "./rootReducer.js";
+
+const store = configureStore({
+  reducer: rootReducer,
+});
+
+export default store;
